requirements.txt
absl-py==2.0.0
ansi2html==1.8.0
arrow==1.2.3
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.2
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==3.3.2
click==8.1.7
cmake==3.27.7
colorama==0.4.6
comm==0.1.4
ConfigArgParse==1.7
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
dash==2.14.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses==0.6
DateTime==5.3
decorator==5.1.1
fastjsonschema==2.19.0
Flask==2.2.5
fonttools==4.25.0
freetype-py==2.4.0
future==0.18.3
google-auth==2.23.4
google-auth-oauthlib==0.4.6
grpcio==1.59.3
idna==3.4
imageio==2.31.2
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipdb==0.13.13
ipython==7.34.0
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
jupyter_core==4.12.0
jupyterlab-widgets==3.0.9
kiwisolver @ file:///C:/b/abs_88mdhvtahm/croot/kiwisolver_1672387921783/work
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib @ file:///C:/b/abs_ae02atcfur/croot/matplotlib-suite_1667356722968/work
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
munkres==1.1.4
nbformat==5.7.0
nest-asyncio==1.5.8
networkx==2.6.3
numpy @ file:///C:/ci/numpy_and_numpy_base_1653574840943/work
oauthlib==3.2.2
open3d==0.17.0
opencv-python==4.8.1.78
openmesh==1.2.1
packaging==23.2
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
plotly==5.18.0
ply==3.11
prompt-toolkit==3.0.41
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyglet==1.5.15
Pygments==2.17.1
PyOpenGL==3.1.0
pyparsing @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_7f_7lba6rl/croots/recipe/pyparsing_1661452540662/work
PyQt5==5.15.7
PyQt5-sip @ file:///C:/Windows/Temp/abs_d7gmd2jg8i/croots/recipe/pyqt-split_1659273064801/work/pyqt_sip
pyrender==0.1.36
pyrsistent==0.19.3
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2023.3.post1
pywin32==306
PyYAML==5.1.1
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rsa==4.9
Rtree==1.0.1
scikit-learn==1.0.2
scipy==1.7.3
sip @ file:///C:/Windows/Temp/abs_b8fxd17m2u/croots/recipe/sip_1659012372737/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smplx==0.1.28
svgwrite==1.4.3
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
times==0.7
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomli==2.0.1
torch==1.7.0
torchaudio==0.7.0
torchgeometry==0.1.2
torchvision==0.8.0
tornado @ file:///C:/ci/tornado_1662476933490/work
tqdm==4.66.1
traitlets==5.9.0
Tree==0.2.4
trimesh==4.0.4
typing_extensions @ file:///C:/Windows/TEMP/abs_dd2d0moa85/croots/recipe/typing_extensions_1659638831135/work
urllib3==2.0.7
wcwidth==0.2.11
Werkzeug==2.2.3
widgetsnbextension==4.0.9
wincertstore==0.2
zipp==3.15.0
zope.interface==6.1
